groq_app/requirements.txt
nemoguardrails==0.14.1
langchain_groq==0.3.6
python-dotenv